+++ common/src/main/java/com/freshdigitable/yttt/data/model/YouTubePlaylist.kt
@@ -36,15 +36,3 @@ interface Updatable {
         fun Updatable.isUpdatable(current: Instant): Boolean = updatableAt <= current
     }
 }
-
-data class YouTubePlaylistItemEntity(
-    override val id: YouTubePlaylistItem.Id,
-    override val playlistId: YouTubePlaylist.Id,
-    override val title: String,
-    override val channel: YouTubeChannel,
-    override val thumbnailUrl: String,
-    override val videoId: YouTubeVideo.Id,
-    override val description: String,
-    override val videoOwnerChannelId: YouTubeChannel.Id?,
-    override val publishedAt: Instant,
-) : YouTubePlaylistItem
