+++ common/src/test/java/com/freshdigitable/yttt/data/model/YouTubePlaylistWithItemsTest.kt
@@ -270,3 +270,15 @@ private fun playlistItem(
     videoOwnerChannelId = null,
     publishedAt = publishedAt,
 )
+
+data class YouTubePlaylistItemEntity(
+    override val id: YouTubePlaylistItem.Id,
+    override val playlistId: YouTubePlaylist.Id,
+    override val title: String,
+    override val channel: YouTubeChannel,
+    override val thumbnailUrl: String,
+    override val videoId: YouTubeVideo.Id,
+    override val description: String,
+    override val videoOwnerChannelId: YouTubeChannel.Id?,
+    override val publishedAt: Instant,
+) : YouTubePlaylistItem
